+++ sdk/inc/azure/core/az_log.h
@@ -28,13 +28,11 @@
 #include <azure/core/_az_cfg_prefix.h>
 
 #define _az_LOG_MAKE_CLASSIFICATION(facility, code) \
-  ((int32_t)((uint32_t)(facility) << 16) | (uint32_t)(code))
+  (((uint32_t)(facility) << 16U) | (uint32_t)(code))
 
 /**
  * @brief Identifies the classifications of log messages produced by the SDK.
  */
-// Make sure to update #_az_log_classifications_are_valid() whenever a new value is added to this
-// enum.
 typedef enum
 {
   AZ_LOG_END_OF_LIST