+++ sdk/src/azure/iot/az_iot_hub_client_twin.c
@@ -108,22 +108,22 @@ AZ_NODISCARD az_result az_iot_hub_client_twin_patch_get_publish_topic(
 AZ_NODISCARD az_result az_iot_hub_client_twin_parse_received_topic(
     az_iot_hub_client const* client,
     az_span received_topic,
-    az_iot_hub_client_twin_response* out_twin_response)
+    az_iot_hub_client_twin_response* out_response)
 {
   _az_PRECONDITION_NOT_NULL(client);
   _az_PRECONDITION_VALID_SPAN(received_topic, 1, false);
-  _az_PRECONDITION_NOT_NULL(out_twin_response);
+  _az_PRECONDITION_NOT_NULL(out_response);
   (void)client;
 
-  az_result result;
+  az_result result = AZ_OK;
 
-  int32_t twin_index;
+  int32_t twin_index = az_span_find(received_topic, az_iot_hub_twin_topic_prefix);
   // Check if is related to twin or not
-  if ((twin_index = az_span_find(received_topic, az_iot_hub_twin_topic_prefix)) >= 0)
+  if (twin_index >= 0)
   {
     _az_LOG_WRITE(AZ_LOG_MQTT_RECEIVED_TOPIC, received_topic);
 
-    int32_t twin_feature_index;
+    int32_t twin_feature_index = -1;
     az_span twin_feature_span
         = az_span_slice(received_topic, twin_index, az_span_size(received_topic));
 
@@ -143,9 +143,9 @@ AZ_NODISCARD az_result az_iot_hub_client_twin_parse_received_topic(
           &index);
 
       // Get status and convert to enum
-      uint32_t status_int;
+      uint32_t status_int = 0;
       _az_RETURN_IF_FAILED(az_span_atou32(status_str, &status_int));
-      out_twin_response->status = (az_iot_status)status_int;
+      out_response->status = (az_iot_status)status_int;
 
       if (index == -1)
       {
@@ -158,20 +158,20 @@ AZ_NODISCARD az_result az_iot_hub_client_twin_parse_received_topic(
       _az_RETURN_IF_FAILED(
           az_iot_message_properties_init(&props, prop_span, az_span_size(prop_span)));
       _az_RETURN_IF_FAILED(az_iot_message_properties_find(
-          &props, az_iot_hub_client_request_id_span, &out_twin_response->request_id));
+          &props, az_iot_hub_client_request_id_span, &out_response->request_id));
 
-      if (out_twin_response->status == AZ_IOT_STATUS_NO_CONTENT)
+      if (out_response->status == AZ_IOT_STATUS_NO_CONTENT)
       {
         // Is a reported prop response
-        out_twin_response->response_type = AZ_IOT_CLIENT_TWIN_RESPONSE_TYPE_REPORTED_PROPERTIES;
+        out_response->response_type = AZ_IOT_HUB_CLIENT_TWIN_RESPONSE_TYPE_REPORTED_PROPERTIES;
         _az_RETURN_IF_FAILED(az_iot_message_properties_find(
-            &props, az_iot_hub_twin_version_prop, &out_twin_response->version));
+            &props, az_iot_hub_twin_version_prop, &out_response->version));
       }
       else
       {
         // Is a twin GET response
-        out_twin_response->response_type = AZ_IOT_CLIENT_TWIN_RESPONSE_TYPE_GET;
-        out_twin_response->version = AZ_SPAN_EMPTY;
+        out_response->response_type = AZ_IOT_HUB_CLIENT_TWIN_RESPONSE_TYPE_GET;
+        out_response->version = AZ_SPAN_EMPTY;
       }
 
       result = AZ_OK;
@@ -190,11 +190,11 @@ AZ_NODISCARD az_result az_iot_hub_client_twin_parse_received_topic(
       _az_RETURN_IF_FAILED(
           az_iot_message_properties_init(&props, prop_span, az_span_size(prop_span)));
       _az_RETURN_IF_FAILED(az_iot_message_properties_find(
-          &props, az_iot_hub_twin_version_prop, &out_twin_response->version));
+          &props, az_iot_hub_twin_version_prop, &out_response->version));
 
-      out_twin_response->response_type = AZ_IOT_CLIENT_TWIN_RESPONSE_TYPE_DESIRED_PROPERTIES;
-      out_twin_response->request_id = AZ_SPAN_EMPTY;
-      out_twin_response->status = AZ_IOT_STATUS_OK;
+      out_response->response_type = AZ_IOT_HUB_CLIENT_TWIN_RESPONSE_TYPE_DESIRED_PROPERTIES;
+      out_response->request_id = AZ_SPAN_EMPTY;
+      out_response->status = AZ_IOT_STATUS_OK;
 
       result = AZ_OK;
     }