+++ sdk/tests/iot/common/test_az_iot_common.c
@@ -244,7 +244,7 @@ static void _log_listener(az_log_classification classification, az_span message)
   }
 }
 
-static void test_az_iot_provisioning_client_logging_succeed()
+static void test_az_iot_calculate_retry_delay_logging_succeed()
 {
   az_log_classification const classifications[] = { AZ_LOG_IOT_RETRY, AZ_LOG_END_OF_LIST };
   az_log_set_classifications(classifications);
@@ -253,6 +253,20 @@ static void test_az_iot_provisioning_client_logging_succeed()
   _log_retry = 0;
   assert_int_equal(2229, az_iot_calculate_retry_delay(5, 1, 500, 100000, 1234));
   assert_int_equal(_az_BUILT_WITH_LOGGING(1, 0), _log_retry);
+
+  az_log_set_callback(NULL);
+  az_log_set_classifications(NULL);
+}
+
+static void test_az_iot_calculate_retry_delay_no_logging_succeed()
+{
+  az_log_classification const classifications[] = { AZ_LOG_END_OF_LIST };
+  az_log_set_classifications(classifications);
+  az_log_set_callback(_log_listener);
+
+  _log_retry = 0;
+  assert_int_equal(2229, az_iot_calculate_retry_delay(5, 1, 500, 100000, 1234));
+  assert_int_equal(_az_BUILT_WITH_LOGGING(0, 0), _log_retry);
 
   az_log_set_callback(NULL);
   az_log_set_classifications(NULL);
@@ -703,7 +717,8 @@ int test_az_iot_common()
     cmocka_unit_test(test_az_iot_status_retriable_translate_success),
     cmocka_unit_test(test_az_iot_calculate_retry_delay_common_timings_success),
     cmocka_unit_test(test_az_iot_calculate_retry_delay_overflow_time_success),
-    cmocka_unit_test(test_az_iot_provisioning_client_logging_succeed),
+    cmocka_unit_test(test_az_iot_calculate_retry_delay_logging_succeed),
+    cmocka_unit_test(test_az_iot_calculate_retry_delay_no_logging_succeed),
     cmocka_unit_test(test_az_span_copy_url_encode_succeed),
     cmocka_unit_test(test_az_span_copy_url_encode_insufficient_size_fail),
     cmocka_unit_test(test_az_iot_message_properties_init_succeed),